feat: add int_taxonomies and refine stg models and constraints

+++ projects/demo/models/staging/decidim/taxonomies/stg_decidim_taxonomizations.sql
@@ -15,10 +15,10 @@
     FROM {{ source('decidim', 'decidim_taxonomizations') }}
 {% else %}
     SELECT
-        CAST(NULL AS INTEGER) AS id,
-        CAST(NULL AS INTEGER) AS taxonomy_id,
+        CAST(NULL AS BIGINT) AS id,
+        CAST(NULL AS BIGINT) AS taxonomy_id,
         CAST(NULL AS TEXT) AS taxonomizable_type,
-        CAST(NULL AS INTEGER) AS taxonomizable_id,
+        CAST(NULL AS BIGINT) AS taxonomizable_id,
         CAST(NULL AS TIMESTAMP) AS created_at,
         CAST(NULL AS TIMESTAMP) AS updated_at
     LIMIT 0
